backend/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Backend for the LangGraph agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.2",
    "google-genai>=1.38.0",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.7",
    "langgraph-api>=0.4.22",
    "langgraph-cli>=0.4.2",
    "langgraph-sdk>=0.2.8",
    "python-dotenv>=1.1.1",
]
